restructure purview models
